+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## Version [1.1.41] - Unauthorized access
+**P.R Raised by** : shubh404-SE on **Date** : 2026-03-06
+### Changed
+  - Updated middleware.ts to:
+    - Decode user role from ck_access_token / ck_refresh_token (JWT payload).
+    - Check permission using existing lib/route-guard.ts.
+    - If role is not allowed for that URL, redirect to dashboard base or login
+---
 ## Version [1.1.40] - ui-layout dashboard
 **P.R Raised by** : shubh404-SE on **Date** : 2026-03-06
 ### Changes